6-add test for method to add user and joke

+++ src/main/java/pl/dawydiuk/repository/RoleRepsitory.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Repository;
 import pl.dawydiuk.domain.Role;
 
 @Repository
-public interface RoleRepsitory extends JpaRepository<Role, Integer> {
+public interface RoleRepsitory extends JpaRepository<Role, Long> {
 
 
     Role findByRole(String role);
